move tests plotting
